GitGud-main/backend/requirements.txt
beautifulsoup4==4.13.4
fastapi==0.115.12
httpx==0.28.1
langchain==0.3.24
langchain_core==0.3.58
langgraph==0.4.1
langgraph_reflection==0.0.1
motor==3.7.0
passlib==1.7.4
pydantic==2.11.4
pydantic_ai==0.1.9
PyJWT==2.10.1
pymongo==4.9
python-dotenv==1.1.0
python_jose==3.4.0
Requests==2.32.3
starlette==0.46.2
duckduckgo_search==7.5.0
bcrypt==3.2.0